chore(api): move dissociation endpoint for admin to src
