--- api/lib/application/organization-learners/index.js
@@ -1,44 +1,0 @@
-import JoiDate from '@joi/date';
-import BaseJoi from 'joi';
-const Joi = BaseJoi.extend(JoiDate);
-
-import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
-import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
-import { organizationLearnerController } from './organization-learner-controller.js';
-
-const register = async function (server) {
-  const adminRoutes = [
-    {
-      method: 'DELETE',
-      path: '/api/admin/organization-learners/{id}/association',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: organizationLearnerController.dissociate,
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationLearnerId,
-          }),
-        },
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
-            '- Elle dissocie un utilisateur d’un prescrit',
-        ],
-        tags: ['api', 'admin', 'organization-learners'],
-      },
-    },
-  ];
-
-  server.route([...adminRoutes]);
-};
-
-const name = 'organization-learners-api';
-export { name, register };
