--- api/lib/application/organization-learners/organization-learner-controller.js
@@ -1,11 +1,0 @@
-import { usecases } from '../../domain/usecases/index.js';
-
-const dissociate = async function (request, h) {
-  const organizationLearnerId = request.params.id;
-  await usecases.dissociateUserFromOrganizationLearner({ organizationLearnerId });
-  return h.response().code(204);
-};
-
-const organizationLearnerController = { dissociate };
-
-export { organizationLearnerController };
