+++ api/lib/routes.js
@@ -11,7 +11,6 @@ import * as certificationPointOfContacts from './application/certification-point
 import * as certifications from './application/certifications/index.js';
 import * as frameworks from './application/frameworks/index.js';
 import * as memberships from './application/memberships/index.js';
-import * as organizationLearners from './application/organization-learners/index.js';
 import * as organizations from './application/organizations/index.js';
 import * as passwords from './application/passwords/index.js';
 import * as poleEmploi from './application/pole-emploi/index.js';
@@ -36,7 +35,6 @@ const routes = [
   certifications,
   healthcheck,
   memberships,
-  organizationLearners,
   organizations,
   passwords,
   poleEmploi,
