+++ api/src/prescription/learner-management/domain/usecases/dissociate-user-from-organization-learner.js
@@ -1,4 +1,4 @@
-import { OrganizationLearnerCannotBeDissociatedError } from '../../../src/shared/domain/errors.js';
+import { OrganizationLearnerCannotBeDissociatedError } from '../../../../../src/shared/domain/errors.js';
 
 const dissociateUserFromOrganizationLearner = async function ({
   organizationLearnerId,
