--- api/tests/acceptance/application/organization-learners/organization-learner-controller_test.js
@@ -1,40 +1,0 @@
-import {
-  createServer,
-  databaseBuilder,
-  expect,
-  generateValidRequestAuthorizationHeader,
-  insertUserWithRoleSuperAdmin,
-} from '../../../test-helper.js';
-
-describe('Acceptance | Controller | organization-learner', function () {
-  let server;
-
-  beforeEach(async function () {
-    server = await createServer();
-  });
-
-  describe('DELETE /api/admin/organization-learners/{id}/association', function () {
-    context('When user has the role SUPER_ADMIN and organization learner can be dissociated', function () {
-      it('should return an 204 status after having successfully dissociated user from organizationLearner', async function () {
-        const organizationId = databaseBuilder.factory.buildOrganization({ isManagingStudents: true }).id;
-        const superAdmin = await insertUserWithRoleSuperAdmin();
-        const userId = databaseBuilder.factory.buildUser().id;
-        const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({ organizationId, userId });
-
-        await databaseBuilder.commit();
-
-        const options = {
-          method: 'DELETE',
-          url: `/api/admin/organization-learners/${organizationLearner.id}/association`,
-          headers: {
-            authorization: generateValidRequestAuthorizationHeader(superAdmin.id),
-          },
-        };
-
-        const response = await server.inject(options);
-
-        expect(response.statusCode).to.equal(204);
-      });
-    });
-  });
-});
