+++ api/tests/prescription/learner-management/acceptance/application/organization-learners-route_test.js
@@ -6,6 +6,7 @@ import {
   databaseBuilder,
   expect,
   generateValidRequestAuthorizationHeader,
+  insertUserWithRoleSuperAdmin,
   knex,
 } from '../../../../test-helper.js';
 
@@ -14,6 +15,31 @@ describe('Acceptance | Prescription | learner management | Application | organiz
 
   beforeEach(async function () {
     server = await createServer();
+  });
+
+  describe('DELETE /api/admin/organization-learners/{id}/association', function () {
+    context('When user has the role SUPER_ADMIN and organization learner can be dissociated', function () {
+      it('should return an 204 status after having successfully dissociated user from organizationLearner', async function () {
+        const organizationId = databaseBuilder.factory.buildOrganization({ isManagingStudents: true }).id;
+        const superAdmin = await insertUserWithRoleSuperAdmin();
+        const userId = databaseBuilder.factory.buildUser().id;
+        const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({ organizationId, userId });
+
+        await databaseBuilder.commit();
+
+        const options = {
+          method: 'DELETE',
+          url: `/api/admin/organization-learners/${organizationLearner.id}/association`,
+          headers: {
+            authorization: generateValidRequestAuthorizationHeader(superAdmin.id),
+          },
+        };
+
+        const response = await server.inject(options);
+
+        expect(response.statusCode).to.equal(204);
+      });
+    });
   });
 
   describe('DELETE /organizations/{id}/organization-learners', function () {
