+++ api/tests/prescription/learner-management/unit/domain/read-models/OrganizationLearnerForAdmin_test.js
@@ -1,6 +1,6 @@
-import { ObjectValidationError } from '../../../../src/shared/domain/errors.js';
-import { OrganizationLearnerForAdmin } from '../../../../src/shared/domain/read-models/OrganizationLearnerForAdmin.js';
-import { expect } from '../../../test-helper.js';
+import { OrganizationLearnerForAdmin } from '../../../../../../src/prescription/learner-management/domain/read-models/OrganizationLearnerForAdmin.js';
+import { ObjectValidationError } from '../../../../../../src/shared/domain/errors.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Read-models | OrganizationLearnerForAdmin', function () {
   describe('#constructor', function () {
