+++ api/tests/prescription/learner-management/unit/domain/usecases/dissociate-user-from-organization-learner_test.js
@@ -1,6 +1,6 @@
-import { usecases } from '../../../../lib/domain/usecases/index.js';
-import { OrganizationLearnerCannotBeDissociatedError } from '../../../../src/shared/domain/errors.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { usecases } from '../../../../../../src/prescription/learner-management/domain/usecases/index.js';
+import { OrganizationLearnerCannotBeDissociatedError } from '../../../../../../src/shared/domain/errors.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | dissociate-user-from-organization-learner', function () {
   const organizationId = 1;
